backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
python-multipart
pydantic-settings
pillow
numpy
mediapipe
opencv-python
